Add (Temporarily duplicated) cleanup routine
As the segmentation algorithm is implemented in C++ and Python, it requires passing data from
C++ to Python several times. Bindings would be the most suitable choice here, but due to ignorance
and time constraints, I have chosen to split the C++ PCL-based binary into two different binaries
temporarily. This will be corrected in the future.